Add separate cache for large decls
Summary: Cachelib has a maximum size for cache items of about 4MB. Larger items can be written using the chained items API, but this isn't ideal for ocamlrep slabs--they need to be stored contiguously. If we never go forward with the zero-copy strategy (i.e., we give up on the idea that slabs in cachelib must be rebased to their current address), this might not be so bad--we can copy the slab pieces into a contiguous slice of memory on the worker side. But it's much easier to do what's in this diff: add a separate cache for large decls. This comes with the cost of needing to send the decl blob over our socket (instead of exposing the blob via cachelib sharedmem), but since there are very few decls exceeding this size, I'm not very concerned about this cost. We can always address it in the future by using the chained item API instead.

Reviewed By: shiqicao

Differential Revision: D23689810

fbshipit-source-id: 132a1b4855a6926dd6d35b6dd3dae8e4a44c15fc