Update the way debugger REPL accesses $this
Summary:
Prior to this diff debugger repl was accessing $this in the autogenerated function. This will no longer be possible with upcoming changes to ban $this in functions/static methods.

This diff renames `$this` to `$__debugger$this` so makes it so that `$this` or `nullptr` is passed in its place.

Reviewed By: paulbiss, dabek

Differential Revision: D23792808

fbshipit-source-id: 6c270b3d844d1267559a9cc5ae36b6c10f6bd16a