Save a function's capability as a fake local and pass it as a parameter to functions
Summary:
This is the meat of the implementation of coeffects as implicit parameters. Before typing a body, we set two non-denotable locals:

- `$#local_capability` - this is exactly `f_cap`. It states what operations are legal in a function body (i.e. `throw`ing, reading from disk, etc) -- currently unused
- `$#capability` - this is the intersection of `f_cap` and `f_unsafe_cap`. This is the capability that is provided as an argument to `ft_implicit_params.capability` for typing function calls, and is the one that will be captured in closures for contextful functions.

Typing of a body proceeds normally. When we reach a call, immediately after checking the arguments, we add a check for all (1) implicit arguments; this is where we read `$#capbility` from the environment. The check is straightforward subtyping, but I discard the unification error message for something more custom.

Here is the example from *rx_example_provisional.php*

```
$ cat /tmp/f.php
<?hh

<<file: __EnableUnstableFeatures('coeffects_provisional', 'union_intersection_type_hints')>>
function f()@{ \HH\Contexts\pure }: void {
  f();
  g();
}
function g()@{ \HH\Contexts\non_rx }: void {
  f();
  g();
}

$ hh_single_type_check /tmp/f.php
Typing[4388] This call is not allowed because its coeffects are incompatible with the context [1]
-> From this declaration, the context of this function body provides the capabilities: mixed [2]
-> But the function being called requires the capabilities: (RxLocal & Throwing<mixed> & NonDet) [3]

/tmp/f.php:6:3
    2 |
    3 | <<file: __EnableUnstableFeatures('coeffects_provisional', 'union_intersection_type_hints')>>
[2] 4 | function f()@{ \HH\Contexts\pure }: void {
    5 |   f();
[1] 6 |   g();
    7 | }
[3] 8 | function g()@{ \HH\Contexts\non_rx }: void {
    9 |   f();
   10 |   g();

1 error found.
```

And in color:

{F336138211}

Reviewed By: losvald

Differential Revision: D22809097

fbshipit-source-id: 38d2c6f0a69e3e6af6b0f9467c00ee58bd5d7ced