Test HHVM runtime prep for "enum classes"
Summary:
Add some unit tests of the changes provided in T75258206.

The syntax introduced by this diff is a placeholder for a syntax that will be introduced in the context of the enum dependent types language feature. Its presence here is to enable early testing of the runtime while EDT syntax remains in development.

Reviewed By: paulbiss

Differential Revision: D23741709

fbshipit-source-id: feed11ec832fd7ce768a2d5464f3e0d628ac2e7a