Make make_token explicitly match in direct decl parser
Summary:
I wanted to reuse the implementation for intersection types in the direct decl parser, which takes 3 items: left paren, type list, right paren. It computes the position between `(` and `)` for the type reason and constructs an intersection from the list. A capability type is exactly this except that instead of `(` and `)` we use `[` and `]`.

The position computation uses the `unwrap_or_return!` macro which early returns `Node::Ignored` if either token is `Ignored`. Unfortunately, because `TokenKind::LeftBracket` wasn't explicitly constructed as a `Node::Pos(...)` like `LeftParen`, `RightParen`, and, curiously, `RightBracket`, it was becoming `Node::Ignored` even before the parser got to calling the `make_capability` smart constructor.

To make things slightly more obvious, I've removed the wildcard and expanded the match to properly cover every token, and I've moved `LeftBracket` up to the set that returns `Node::Pos(...)`

Differential Revision: D23800343

fbshipit-source-id: 4adb9778a1e05d8a47c4d52750b63fa3ca1d312a