Add --type-at-pos-batch-ex and thread it through
Summary:
I have an idea for a new implementation of --type-at-pos-batch which should deliver better perf in the egregious cases, and should be about equivalent in other cases.

This diff plumbs through the switch. My design is that `--type-at-pos-batch` will use the existing implementation, and `--type-at-pos-batch-ex` will use my new experimental implementation.

Reviewed By: hgoldstein

Differential Revision: D23740238

fbshipit-source-id: ce94fd8db5ac377eef047406394f4a2520b99c89