Fix copy-on-write
Summary:
Introduces a dedicated `Tcow_array` policy type that will be used for Hack arrays i.e. `vec`, `dict`, and `keyset`. It holds the value policy type and policies for key. Both of them are covariant.

Covariant treatment of its constitutent policy types revert the imprecision introduced by `Tclass` simplification and brings back precise copy-on-write semantics (hence the changes in the check tests).

This diff does not deal with the constraints due to indexing and length. That'll come soon in a future diff.

Reviewed By: zilberstein, mpu

Differential Revision: D23783065

fbshipit-source-id: 4aa1defcdf325eca9d5e409d02bf556500901636