fb_setprofile(): do not report arguments, $this and reified generics when resuming
Summary:
From the profiler perspective, it makes sense to obtain the list of passed
arguments, `$this` and reified generics when a function is being called, but
there's not much reason to do so when resuming a previously suspended async
function or generator. We don't report these fields for other events either,
such as returning, suspending or throwing an exception.

Reviewed By: ricklavoie, paulbiss

Differential Revision: D23810336

fbshipit-source-id: a7f8aa923fd4f19b1df62a0e9d6f52b03473e4db