Specify the visitor type in the desugared function
Summary:
When we encounter an expression tree:

```
$c = Foo`...`;
```

ensure that the desugared visitor expression specifies the visitor type as a parameter.

```
$c_desugar = (Foo $v) ==> $v->...`;
```

This gives us more type checking over the previous `dynamic` type hint.

Also ensure that the desugared visitor expression uses positions from the original user syntax. This enables us to highlight the relevant part of the expression tree which is using unsupported syntax for the given visitor.

Differential Revision: D23776180

fbshipit-source-id: 7599cd4679e4441ca4a409d7cc2c05ede3f1bed6