Add support for renaming TypeAlias's in VSCode
Summary:
I was trying to rename a type alias via VSCode, but got the error message "Tried to rename a non-renamble symbol"

After that, I tried to rename it like so `hh --refactor Class IGPaymentPinCreateResponsePayload IGPaymentPinCreateResponsePayloadDEPRECATED` and it worked.

I was curious as to why the VSCode one might not work, so I grepped for th error message and figured I'd take a stab at fixing it.

Reviewed By: DavidSnider

Differential Revision: D23873001

fbshipit-source-id: 53085acb09c6240059c92450e1271b18814a8091