Eliminate DataTypeCountness
Summary:
DataTypeCountness is a strange guard-constraint level. It doesn't actually imply a logical constraint, but it still causes us to split translations on DataType, and then we may later use profiling to sometimes merge these translations (except that we'll still keep counted DataTypes separate). Our logic has several special cases for this constraint level.

We can eliminate this constraint without hurting performance - in fact, it looks like a small win. We use this constraint level in those cases when all we're going to do with a value is dec-ref it. In these cases, it's generally better to load a generic value and let DecRef code layout choose the right one. Consider the following examples, where we have a value with a given distribution, a DataTypeCountness constraint, and we do a DecRef:

| Value distribution | Old codegen | New codegen
| 95% int, 5% object |  Guard relaxation kicks in; test for uncounted DataType, then test for object, and do the dec-ref. | Single DecRef, optimized for the "kind tells us we're done" case. Comparable code to before; we replace a type test with a refcount check.
| 30% int, 30% bool, 10% object, 10% vec, 10% dict, 10% keyset | Guard relaxation kicks in; test for uncounted DataType first, then test for each of the 4 counted types. For the array-like ones, we still need to check the sign of the refcount before doing the dec-ref. | Single DecRef, optimized for the equiprobable case. Better code than before; no type tests, and equally good dec-ref.
| 10% of each counted Hack type | Guard relaxation does not kick in. 10 type tests and 10 DecRefs. | Single DecRef, optimized for the counted case. Much better code than before; no type tests and much smaller code size.

Reviewed By: ricklavoie, ottoni

Differential Revision: D23742044

fbshipit-source-id: 3e0e14065a5a24afcb9769b826376a31d96306b0