Use 63-bit hashes if custom dependency graph is initialized
Summary:
In the previous diff we refactored `Typing_deps.ml` to actually query the 64-bit dependency graph via Rust. However, we were still generating 32-bit hashes to do so!

This diff fixes that: **generate 64-bit hashes when the 64-bit dependency graph is initialized!**

The SQLite naming table is unchanged: we still use a combination of the original 31-bit dependency hash, and the 31-bit naming hash. Making this work requires a small fix in querying, as we need to truncate 63-bit hashes to 31-bit hashes (and not 32-bit hashes!).

Differential Revision: D23396982

fbshipit-source-id: c3ef7e58c99dc0c35a95fa0c9a5e834e845a654f