Add permissive Pos::merge helper
Summary:
We frequently use the `unwrap_or_return!` macro when fetching or merging positions. The macro takes an `Option<T>` and unwraps it to `T` if it's `Some`, or returns `Node::Ignored` otherwise. The use of this macro is usually a code smell--it indicates a place where our error recovery is poor (just because some token in a function declaration is missing does not mean that we should behave as if the entire function declaration was missing).

Since Pos has a None value already, we could replace the `Option<Pos>` values on which we are using `unwrap_or_return!` with just a `Pos` (which may be `Pos::none`, if the node we fetched the position from happened to be missing). This diff does so by adding the helper methods `merge` (which merges two optional positions into a non-optional position which may be `Pos::none`) and `merge_positions` (which takes two `Node`s, fetches their positions, and `merge`s them).

This strategy carries some risk. It is now easier to accidentally write `Pos::none` into the decl we parse from correct code (i.e., there is no guarantee that our `Pos::none` behavior only comes up during error recovery). We may want to add a test that ensures that all positions are non-`none` in the decls of files without errors.

This diff removes half of our uses of `unwrap_or_return!`, bringing the count from ~160 down to 80.

Reviewed By: vassilmladenov

Differential Revision: D23909624

fbshipit-source-id: 22841a5776637a60eafb933e34628deae0dc312e