Return Pos from get_pos instead of Option<Pos>
Summary:
> We frequently use the `unwrap_or_return!` macro when fetching or merging positions. The macro takes an `Option<T>` and unwraps it to `T` if it's `Some`, or returns `Node::Ignored` otherwise. The use of this macro is usually a code smell--it indicates a place where our error recovery is poor (just because some token in a function declaration is missing does not mean that we should behave as if the entire function declaration was missing).

A few of our uses of `unwrap_or_return!` are for the result of `get_pos`, which returns `None` when the given node does not contain a position. I think these are all instances of poor error recovery--it makes more sense to use `Pos::none` when an expected node happens to not be present (especially considering the strategy implemented in the previous diff).

This diff also moves get_pos from being a method on Node to a method on DirectDeclSmartConstructos, so that it has access to `self.state.arena` rather than needing to take it as a parameter.

Differential Revision: D23910808

fbshipit-source-id: 0c4b1d42f6bb5dc87b107c5791327358aac55993