Replace special casing for shape keys by subtype against arraykey
Summary:
In a shape type or shape expression, we currently check that the keys have types that are `int` or `string` (or an enum) by an adhoc case analysis over the type that has to deal with intersections etc. Also, we break through any opaque `newtype` to find the underlying type. This goes against the spirit of `newtype`!

Instead, let's replace with a simple subtype check against `arraykey`, letting the subtype logic deal with cases such as intersections. An arraykey is in the end an int or a string, or an enum. If this should change in future, we will want to rethink this check, but that also goes for many uses of arraykey in our type system.

Another advantage of checking against `arraykey` is that fewer dependencies will be generated: a quick run of `hh_single_type_check --dump-deps` on some of the test cases reveals that we end up transitively depending on the whole chain of `newtype` definitions. This isn't necessary for simple `arraykey` checks.

Reviewed By: vsiles

Differential Revision: D23904115

fbshipit-source-id: ebdd7c6f944e4a4be4409166ce000a21140e9b29