OutOfBoundsException due to Hack arrays
Summary:
Reading non-existent keys from `dict`s and indices from `vec` as well as writing to non-existent indices for `vec`s lead to `OutOfBoundsException`s. This diff implements the flows that happen due to these exceptions.

It also adds a policy to be used for `length`, but does not implement all constraints related to array length just enough to handle `OutOfBoundsExceptions` i.e. flow to the exception from pc.

Reviewed By: zilberstein

Differential Revision: D23816468

fbshipit-source-id: d66c6ffed3b8b45b67693251145805cbe5737e80