don't sort alphabetically
Summary:
There are several decl accessors which sort their list of members alphabetically.

(I had thought that this sorting was breaking their laziness. But in fact, the way we implement it is by eagerly building the full LSTable, and then turning the list of members into a sequence. Thus, converting the sequence back into a list and sorting it doesn't break laziness. I think basically that anytime we invoke an API called "get all consts" then we're pretty explicitly being eager!)

I figure there's no good reason to sort a list of members: our typechecking should be invariant whether a method is named "aardvark" or "zodiac".

The only possible place that sorting should be apparent is in error lists. And then, it's a waste of CPU time for us to be sorting unless we actually display errors.

In cases like LSP, the error list is actually meant to be unsorted, since sorting happens on the client. The only place that can possibly have an order-sensitive error list is command-line error reporting.

And in that case, we should be sorting the errors after they're generated. (indeed, with distributed typechecking, we can't expect a stable order of errors unless they're sorted).

THEREFORE

* This diff changes typing_classes_heap.ml to not sort its member list
* It does minor tweaks in autocompleteService and switch_check, where it makes reasonable sense to sort things
* It resets the "expected" order for just two files in our test suite which are sensitive to the order in which methods are typechecked

Reviewed By: shiqicao

Differential Revision: D23698327

fbshipit-source-id: be9fe60e21f1f231c4c3dfbceb2a735b7ad70087