Eager classes now always read from shmem
Summary:
It used to be that users of eager (folded) decls would always write to the decl heap, but if using the Local_memory_provider then it didn't read from the decl heap.

This doesn't make sense to me. I can't imagine why it's useful. So, I simplified the code by having eager decls always read from the decl heap as well as write.

Differential Revision: D23697473

fbshipit-source-id: d396d0f35d69d704a6f3f68a00709aa6b7806c0e