Bespoke profiling codegen improvements
Summary:
Recently, we've observed a regression when bespokes are enabled. Our working hypothesis is that this is partially due to profiling tracelets being slower and collecting less data. We address a few inefficiencies here:
1) We check whether we have a bespoke array before calling LogArrayReach to avoid the native call if we have a vanilla array.
2) We mark the interp side of diamonds as unlikely.

Note: there is an explicit check for logging arrays in the native helper `arrayReach` so that this doesn't break when it's possible for other bespokes to reach profiling tracelets. When we guard profiling tracelets on {Logging|Vanilla}, this will no longer be necessary.

Differential Revision: D23915423

fbshipit-source-id: c5e7b567141fed9fc5d67c1e4c656b05d8171eb7