Return Attributes instead of Option<Attributes> from to_attributes
Summary:
> We frequently use the `unwrap_or_return!` macro when fetching or merging positions. The macro takes an `Option<T>` and unwraps it to `T` if it's `Some`, or returns `Node::Ignored` otherwise. The use of this macro is usually a code smell--it indicates a place where our error recovery is poor (e.g., just because some token in a function declaration is missing does not mean that we should behave as if the entire function declaration was missing).

In our helper `Node::as_attributes`, we return `Option<Attributes>`, but this seems to be unnecessary--we always return `Some`. Change it to return `Attributes` instead, so we can remove some `unwrap_or_return!` uses.

This diff also moves the method to be a method of `DirectDeclSmartConstructors`, so that it need not take an arena as a parameter.

Reviewed By: vassilmladenov

Differential Revision: D23932754

fbshipit-source-id: 1127f48330348e9b4e108d6f6c18d45709542fed