Include a SyntaxKind on every ignored node
Summary:
The previous diff marks every ignored node returned by `zero` and `flatten` with its syntax kind, to aid in debugging by making it clear which smart constructors function produced the ignored node. This diff extends that to *every* Ignored node, so that it is always possible to tell where an ignored node came from.

This should have no effect on performance because it does not change the size of `Node::Ignored` specifically or `Node` as a whole. Because Rust lays out all enum variants with the same size, `Ignored` was eight bytes of tag data and eight bytes of empty space (because the largest variants have an eight-byte-aligned tag and an eight-byte pointer).
Including a SyntaxKind (as `IgnoredSyntaxKind` did) means the variant is eight bytes of tag data and eight bytes of syntax kind data (SyntaxKind is two bytes, but `IgnoredSyntaxKind`'s data section needs to have the same alignment as the largest variant's data section, which, since it's a pointer, is eight bytes).

Reviewed By: vassilmladenov

Differential Revision: D23932751

fbshipit-source-id: 48e8780283df11f81a147764fe601368e22f2389