Remove unnecessary vec![]
Summary:
We do need to allocate (because we can't return a reference to `self` when we have `self` by value), but there's no need to allocate a resizable vector (with some capacity which may be more than 1)--we can just allocate enough space for a single node (two words) and use `slice::from_ref`.

Also condenses the first two arms of the match, which can be an or-pattern.

Reviewed By: vassilmladenov

Differential Revision: D23933525

fbshipit-source-id: a4bc8dec40c3e393ee229f4119aaca621238665f