Remove maybe_slice_from_iter
Summary:
> We frequently use the `unwrap_or_return!` macro when fetching or merging positions. The macro takes an `Option<T>` and unwraps it to `T` if it's `Some`, or returns `Node::Ignored` otherwise. The use of this macro is usually a code smell--it indicates a place where our error recovery is poor (e.g., just because some token in a function declaration is missing does not mean that we should behave as if the entire function declaration was missing).

The helper `maybe_slice_from_iter` returns `Some(slice)` if all elements in the iterator are `Some`, or `None` otherwise. We often use it with `unwrap_or_return!` to bail if any element in an iterator is not what's expected. This seems like poor error recovery--we could just omit the single unexpected element with a `filter_map` instead.

Reviewed By: vassilmladenov

Differential Revision: D23932752

fbshipit-source-id: bfe39fc1215bec7856c1e35e53cba862f7e4a8d0