Only update concurrent map in shouldTranslateNoSizeLimit when it will be used
Summary:
`shouldTranslateNoSizeLimit` is 0.21% gCPU, but does very little besides looking up a SrcKey's function and updating the function and SrcKey's atomic counters. The SrcKey's atomic counter is not used unless Eval.JitSrcKeyThreshold is enabled, which it does not appear to be in our configurations. If it's not enabled, we can skip the update.

This appears to be a clear memory win.

https://fburl.com/strobelight/7856vg0l

Differential Revision: D23931618

fbshipit-source-id: f5871b3dddbc2baac22b1ffddbf0a13661576ff0