Harden reified typehints post-specialization
Summary:
When we finished HAM specialization, we modified "is shape" checks to restrict to darray inputs and "is tuple" checks to restrict to varrays, but the same restriction didn't apply to reified generics typehints. Today, a varray will pass a shape reified generics typehint and a darray will pass a tuple reified generics typehint.

Reified generics typehints should be checked at the same granularity as any other param, prop, or return typehint; that is, we should check the top-level type, but not the inner type.

Reviewed By: ricklavoie

Differential Revision: D23866668

fbshipit-source-id: f99f4f72da101ae1bc5107c4150de0b13a6b24ed