Properly save/restore dependency graph mode in workers
Summary:
Keeping track of the dependency graph delta, requires us to check whether discovered edges are already present in the graph.

Workers thus need access to the dependency graph. This diffs registers the path to the dependency graph in each worker, allowing each worker to lazily open the dependency graph if needed.

Differential Revision: D23573647

fbshipit-source-id: 3909d2bf8e79ded4a112dcf3e23e11e402b166fc