Legacy bit on arrays option for thrift binary/compact deserializers
Summary:
As part of hack array migration, there are times when thrift deserializers will produce structs with varray and darrays in them, and those flow into locations that need to preserve legacy behaviors. For performance reasons, it's easier to set the legacy bit at the source when these arrays are instantiated than to recursively coerce these arrays in userland on the deserialized struct.

This diff adds a flag that can be threaded to binary/compact deserializers that will set the legacy bit for varrays and darrays created.

Reviewed By: dneiter

Differential Revision: D23919637

fbshipit-source-id: 88f2c17ae03d085b504f44ed47745769d6bd7905