LSTable.to_list
Summary:
This is a trivial cleanup diff. We had several places which (1) eagerly traverse the LSTable, (2) turn it into a list, (3) turn it into a sequence, (4) turn it back into a list.

So I removed LSTable.to_seq and replaced it with LSTable.to_list.

I changed two APIs in Typing_classes_heap.Api. They used to be eager but misleadingly returned a Sequence.t. Now they just return a list. It feels cleaner this way. It has a principle "If a method returns Sequence.t then you safely know it's really lazy. If it returns a list then you safely know it's really eager."

Reviewed By: shiqicao

Differential Revision: D23892173

fbshipit-source-id: c54ad884994061f1d7a4df4497e6b534b62c6bc0