Memcpy decl slabs into OCaml heap
Summary:
We currently copy slabs from cachelib into the OCaml heap using a traversal of the value tree. We start at the root value, allocate an equivalent OCaml value using ocamlpool, then recurse for each of its fields (using `ocamlrep::Value::clone_with_allocator`).

This works okay most of the time, but it breaks down for very long lists, where the lack of tail recursion in `clone_with_allocator` results in a stack overflow. It's also needlessly expensive--a copy-and-rebase of the slab should be sufficient (i.e., a memcpy followed by a linear traversal over the slab to fix up its internal pointers). This diff implements that strategy, and uses it in hh_worker's decl_ipc_ffi.

Reviewed By: shiqicao

Differential Revision: D23968668

fbshipit-source-id: ce542e758e7268731cefc0d4e6873377657c66fb