Decl-parse the file being typechecked before typechecking it
Summary:
In the associated task, we recorded extremely long typecheck times for some files. This seems to be caused by evictions of decls declared within the file being typechecked (since the problematic files are ones which declare thousands of classes). We request the class decls from the decl service as we encounter them. Each time the decl service has a cache miss, it re-parses the entire file and inserts all of its decls into the cache. As we typecheck the file, decls defined later in the file are evicted, and when we reach them for typechecking, the decl service must re-parse the entire file again. It seems that this happens so frequently for some of these long files that they can take minutes to typecheck instead of seconds.

This diff addresses the problem by performing the decl-parse of the file being typechecked in hh_worker. We insert the decls in the Decl_service_client cache, which never evicts them, guaranteeing that the decls in the current file will be available for the duration of the typecheck.

Differential Revision: D23966150

fbshipit-source-id: 75dcd17331c9611edf9a469c071858215acd37ad