Change type parameter mismatch error message
Summary:
Previously, when we gave two type parameters for a function that only expected one, we gave the error: `Expected a type parameter`. However, due to highlighting and wording, this was confusing, as users might think that they already gave a type parameter (they've given two!). Thus change, the wording to `Expected exactly one type parameter`.

Also updated the error on 0 type parameters to use the proper plural form.

Reviewed By: periodic1236

Differential Revision: D23994448

fbshipit-source-id: 907a4e00416f35c07ed6a53cdc915d23b187d66d