Goodbye ArrayData::Merge and Renumber
Summary:
Kill two gross operations on array-likes.

merge() is used in three places: to implement the array_merge builtin, in ext_soap, and in reified generics. Much of the complexity for this op relies in its crazy behavior around integer keys. This behavior can be treated much more simply if we have constraints on the keys of the two arrays being merged:
- If both arrays have vector data, merge is equivalent to concatenation - see ext_soap.
- If both arrays have string keys, merge is equivalent to simply overwriting keys - see reified generics.

Note, in ext_soap, we previously constructed a darray (merge always returns a darray), but we now keep the headers in a varray. That doesn't matter, because all that happens to the headers list downstream is that we iterate over them.

That leaves the (arbitrary, unconstrained) usage in ext_array. In this place, we are always constructing a new darray and merging values into it, so we can assert that the base is vanilla and emit static calls to the MixedArray helpers. That allows us to eliminate the generic merge operation from ArrayData. After that's done, renumber() is dead - the last usage is in the dead function ArrayCommon::Dequeue. (We may bring back some more ArrayCommon functions, but we won't implement them in terms of this problematic primitive.)

In this way, we've not only cleaned up our code, but we've also made all three sites perform better (since they now run specialized alternatives instead).

Reviewed By: ricklavoie

Differential Revision: D23997697

fbshipit-source-id: 7eb3f8700b3734e877cf8cc9ed185ab0da61c09c