var_dump includes dvarrays by default
Summary:
When this flag is on, we print varrays as "varray(...) { ... }" and we don't print keys, and we print darrays as "darray(...) { ... }", analogous to vecs and dicts.

Since var_dump is routed straight to stdout, it shouldn't be used for logic, only debugging. (Someone may collect and parse a Hack file's output, and those people would need to update some code, but I expect this case is rare.)

Reviewed By: viratyosin

Differential Revision: D23998754

fbshipit-source-id: 9ef476e9019855b776755455316e4d15301b665e