simplify code for logging profiling info
Summary:
This diff aims at reducing the noise around the calls to `process_file` (which is the main logic and should be visible to readers) by just having one-liners before and after to read all counters. To do that, we add a `read_counters` function which reads the custom counters defined in `counters.ml` but also time and memory "counters". `read_counters` returns a `Telemetry.t` so we can later use `Telemetry.diff` to get the costs of `process_file`, which allows to kill `diff_mem_telemetry` and `diff_counters_telemetry` and other logic. The produced JSON is slightly reorganized and looks something like:
```
{
  filesize:116350,
  deferment:{
    times_checked:1,
    files_to_declare:168
  },
  profile: {
    time: 6.6957879066; // formerly "duration_decl_and_typecheck"
    memory: {
      gc: ...
      shmem: ...
    };
    operations: {
      decl_accessors: {
        count:13510141,
        time:1.50311493874
      };
      disk_cat: {
        count:13510141,
        time:1.50311493874
      };
      get_ast: {
        count:13510141,
        time:1.50311493874
      };
    };
  }
  // optionally
  profile_second_run: {
    time: ...; // formerly "duration_typecheck"
    memory: { ... };
    operations: {
      decl_accessors: ...;
      disk_cat: ...;
      ...
    };
  }
}
```
One advantage of this renaming is that we are not affirming anymore that the second run is pure typechecking time (as this might not always be the case) while at the same time making it explicit what methodology is being used.

NB: Previously, it used to produce JSONs like
```
{
  "duration_decl_and_typecheck":6.69578790665,
  "duration_typecheck":null, // duration of second typecheck
  "filesize":116350,
  "deferment":{
    "times_checked":1,
    "files_to_declare":168
  },
  "counters":{
    "decl_and_typecheck":{
      "decl_accessors":{
        "count":13510141,
        "time":1.50311493874
      },
      "disk_cat":{
        "count":60,
        "time":0.00620293617249
      },
      "get_ast":{
        "count":65,
        "time":0.150580406189
      }
    }
    ("second_typecheck": ...)
  },
  "mem":{
    "decl_and_typecheck": {
      "gc": ...
      "shmem": ...
    }
    ("second_typecheck"):...
  }
}

Differential Revision: D21740142

fbshipit-source-id: dc366cfffbbebbebf0cba498fbc3851a8db02e0d