Spill uninits over uninitialized ActRec and inouts in prologue's stublogue-mode catches
Summary:
If a prologue throws an exception while in a stublogue mode, we currently sync
the registers using a fixup of the caller, which skips all inputs and also the
reserved space for ActRecs and inouts.

This is problematic for a number of reasons, but before we can fix the fixup,
we need to make sure the stack does not contain uninitialized garbage so that
we can unwind it.

So, let's write uninits over the potentially uninitialized space and do it in
a way that avoids code size increase for the common case.

Reviewed By: ricklavoie

Differential Revision: D24004460

fbshipit-source-id: f0646ff8211d4b95b51eaec7bf8e94874bf1ead6