Define stack outside of IRGS's constructor
Summary:
Stack definition is context specific. Currently, we get a bit whether we are
creating translation for a prologue and we also get initSpOffset. This will not
be sufficient for prologues. Instead of adding more stuff to the TransContext
that is specific only for a single context type, split the logic and move the
stack initialization to two separate methods. One for prologues and another for
everything else.

Reviewed By: ricklavoie

Differential Revision: D24002577

fbshipit-source-id: 11c16678be20d924c841e061e6a808f64f4c7ce5