Explicitly write error codes, remove deprecated ones
Summary:
This diff implements the plan in T76168136.

- Add explicit error codes to every error in error_codes.ml
- Comment out all deprecated error codes.

This diff does NOT delete error_map.ml just yet, I want to use the test one last time to confirm that no other error codes have changed due to this change. The diff on top of this will delete the error_map unit test.

At the end of this stack, the flow for adding an error will be:

- Add a line to error_codes.ml with the new error name, with a corresponding error code:

  ```
       | CallCoeffects [value 4390]
  +    | MyNewErrorCode [value 4391] (* EXTEND HERE WITH NEW VALUES IF NEEDED *)
  ```
- Add a function to errors.ml with the function calling Typing.err_code with the new error(same as before)
- Regenerate rust file with `hphp/hack/src/oxidized/regen.sh` (Same as before) (Getting rid of this step essentially means putting the rust generation into the build instead of manually generating, we may want to do this later as well)

This is better than before, where we would need to update both error_codes.ml and the error_map.ml unit test, needing to change two different files. Moreover, this allows us to actually delete/comment out the deprecated error codes from existence.

Reviewed By: vsiles

Differential Revision: D23967429

fbshipit-source-id: 28b05e201fc6fcd343b268ec39d973c6badc7946