Materialize virtual dispatch structs
Summary:
In order to emit a direct call for a known layout into the TC, we need to create symbols with the signature "TypedValue GetStr(ArrayData*, StringData*)" for each layout class.

One approach to do that is in D23635063. But we can also eliminate these methods from bespoke::Layout completely, evolving the layout closer to the "BespokeSpec" we now know we need, and eliminating some boilerplate along the way. As for the use of macros here: none of them leak outside the "bespoke" directory, and it saves us a lot of unnecessary repetition when declaring these functions for each layout.

Reviewed By: jjgriego

Differential Revision: D24023113

fbshipit-source-id: e68dbc5f6ccb2a77520b69a93e817da32609e4a5