Make all conversions generic
Summary:
We replace the five conversion methods with two methods: ToDVArray and ToHackArr. These methods cover the only cases where we need layout-specialized logic for conversion; all other cases involve a change of layout, so generic code is fine.

In particular, we now have to do all of this logging and special case work for varray->vec and vec->varray conversions. We can do all of that generically, in array-data.cpp, and only call out to virtual methods to actually do the conversion.

Reviewed By: jjgriego

Differential Revision: D24011105

fbshipit-source-id: 2cf4374256c16148e33bf2ea0f230e3ee88b417b