Treat nontrivial casts as constructors
Summary: Any cast operation that is a true change of layout will return a vanilla array. We can take advantage of that for bespoke optimizations. Also, we need to treat dict-cast-on-vec as an array creation site for logging purposes (which we were previously doing in LoggingArray, but we can now do like any other bytecode).

Differential Revision: D24040348

fbshipit-source-id: 2fbc0e07e00899cc53af134876920cc86e9b42cf