Add a StrKeyTable to empty dicts/darrays
Summary:
I noticed that the "missing key" optimization if often not kicking in because of arrays that don't have the StrKeyTable. Looking more into why that was happening, I noticed that it was happening even for some static arrays, and then I found that it was because empty arrays didn't have the StrKeyTable. This diff adds such table to them.

With this change, the "missing key" optimization appears to kick in 2-2.5x more often among the top 100 translations in prod.

Reviewed By: oulgen

Differential Revision: D23924632

fbshipit-source-id: 65332f1f29bc068b2c91f35b5d9726ab7cd24692