Introduce constrained-lval helper
Summary:
We can use this new helper to implement lval operations for monotype array-likes of (counted) array-likes, ints, or objects without escalation. The new helper works with Dim and IncDec member ops; since SetOp can perform arbitrary updates on the type of the value, we're unable to use it there.

We also fix a small bug in entry types tracking: if we append a persistent string to an array, the old code would have preserved KindOfPersistentString as the type; the new code correctly pessimizes the type at that point to KindOfString. That's one option for handling the behavior; another is to always pessimize to KindOfString to begin with. We can play around with it.

Differential Revision: D23942796

fbshipit-source-id: ee9dcc59211a5e4ad404cf669f2a03b7d4cf8352