Defensive telemetry in case of bugs
Summary:
Last time I tried to release type-at-pos-batch-ex, it caused a SEV. hgoldstein advised me to put in more telemetry as a defense against that. Here it is.

This diff makes it log *before* type-at-pos-batch, as well as after. I imagine we could plot a graph of how many Befores/hour and how many Afters/hour. If we get say 95% as many Afters/hour as Befores, that will indicate that 5% of requests are crashing.

Reviewed By: hgoldstein

Differential Revision: D24063088

fbshipit-source-id: 9ebec73d2a32e8c232815a4d9bab29bea95504d9