Desugar continue and break
Summary:
Currently, `break` and `continue` statements are not supported within Expression Trees. We support `while` loops current and plan to support `for` loops, so I think it would be good to support `continue` and `break` as well.

Before the changes in this diff, we would encounter:
```
Code`() ==> {
  while(true) {
    continue;
    // 4 * 4;
  }
}`;

Parsing[1002] Unsupported syntax for expression trees.
```

Now, this diff implements desugaring for `continue;` and `break;` as:

```
$v->continueStatement()
```
and
```
$v->breakStatement()
```
respectively

Reviewed By: shayne-fletcher

Differential Revision: D23945132

fbshipit-source-id: 883e87010cf31e8edb8dc9617fed1ad0a9261bca