Untangle `Typing_object_get` and make its nullsafe behaviour slightly more general
Summary:
Attempt to make the inter-function dependencies clearer in `Typing_object_get` by separating functions into their own `let [rec]` declarations where possible.

Also, allow callers of `Typing_object_get` to pass in a typing reason (instead of just a position) as the `nullsafe` argument. This permits consideration of more general null-safe behaviour, rather than just null-safe method invocations (`?->`).

For example, when checking subtyping for `_ <: Thas_member(...) | null`, this is effectively a "null-safe" method invocation. However, the reason it is null-safe may not be a `?->` - instead it could arise from something like `if ($receiver is nonnull) { $receiver->foo() }`, in which case we would have `_ & nonnull <: Thas_member(...)`, which is transformed to the form above. In this case, the reason for the null-safety is the reason encoded in the `nonnull` type.

Reviewed By: andrewjkennedy

Differential Revision: D23908084

fbshipit-source-id: 29519e83649ec6688300fb5aafa70d9d976915bd